Refactor: open with_* API
Now it's possible to execute something in the context of writer or
reader, meaning that the data will be locked only for the current thread
and user can decide to hold the lock as long desirable.

Add proper randomization for the test data and make sure to test new API
with batch access.